requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
Django==3.1.4
django-multisite==1.9.0
docopt==0.6.2
ExifRead==2.3.2
filelock==3.0.12
filesystemlib==0.0.4
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==8.0.1
pipreqs==0.4.10
pluggy==0.13.1
pycparser==2.20
pytz==2020.4
requests==2.25.1
requests-file==1.5.1
SQLAlchemy==1.3.19
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.1
yarg==0.1.9

